Oof ouch owie rebase-hurting juice :(
I am updating the paramiko version

This will make SSH work. This has already been fixed in the upstream, but I
don't want to rebase yet. If you want to check out the commit so you can
cringe about the cycles I wasted, here's the link:

https://github.com/scoringengine/scoringengine/commit/a4009a4b4a4eaefe3b8cd82b85fd4fecd51248d7

I can't wait for the rebase. It's going to hurt so much.